\name{dot-PBSwickEnv}
\docType{data}
\alias{.PBSwickEnv}
\title{PBSwicked Environment}
\description{
  An environment set aside for PBSwicked.
}
\usage{.PBSwickEnv}
\format{
A new environment with a \code{.GlobalEnv} parent.
}
\source{
  Generated by a call to the base function \code{new.env()}.
}
\details{
  The environment is created in \code{'zzz.r'} and can be
  used by \code{PBSmodelling} functions \code{'lisp'}, \code{'tget'}, 
  \code{'tput'}, \code{'tprint'}, and \code{'tcall'}, if this package has be loaded.
}
\seealso{
In \pkg{PBSmodelling}:\cr
  \code{lisp},
  \code{tget},
  \code{packList}
}
\keyword{environment}

