\name{wickiverb}
\alias{wickiverb}
\title{wickiverb in PBSwicked}
\description{
  \code{wickiverb} is a named character vector containing 26 verb
  choices (delimited by \code{'|'}) in a wicked world. 
  The names of the vector are alphabetic variations on the word \code{wick}.
}
\usage{
wickiverb
}
\format{
  A character vector of length 26.
}
\keyword{datasets}

